rename `home` folder into `discover`
