+++ app/src/main/java/dev/bruno/wheretowatch/features/discover/HomePresenter.kt
@@ -1,4 +1,4 @@
-package dev.bruno.wheretowatch.features.home
+package dev.bruno.wheretowatch.features.discover
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
