+++ app/src/main/java/dev/bruno/wheretowatch/features/discover/movies/UpcomingMovieFlowSource.kt
@@ -1,6 +1,6 @@
-package dev.bruno.wheretowatch.features.home.movies
+package dev.bruno.wheretowatch.features.discover.movies
 
-import dev.bruno.wheretowatch.features.home.HomeMovieItem
+import dev.bruno.wheretowatch.features.discover.HomeMovieItem
 import dev.bruno.wheretowatch.services.discover.DiscoverCategory
 import dev.bruno.wheretowatch.services.discover.DiscoverContentSupplier
 import kotlinx.collections.immutable.ImmutableList
